requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: CLIP model
# Install with: pip install git+https://github.com/openai/CLIP.git
# clip @ git+https://github.com/openai/CLIP.git

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
